Renamed useStorage to useSessionStorage
