+++ src/pages/calendar/index.tsx
@@ -8,7 +8,7 @@ import { Header, Content, Feedback } from '../../app/index';
 import { getCurrentTerm } from '../../app/shared/utils/terms';
 import { ContentSidebar } from '../../app/sidebar';
 import { Term } from '../../shared/fetchers';
-import { useSessionStorage } from '../../shared/hooks/storage/useStorage';
+import { useSessionStorage } from '../../shared/hooks/storage/useSessionStorage';
 
 export function Calendar(): JSX.Element {
   const navigate = useNavigate();
